idk man

+++ redis/entities/timeControl.ts
@@ -6,13 +6,14 @@ interface TimeControl {
     name: string,
     cooldown: Date,
 }
-type TimeTable = Map<string, () => Promise<any>>
-class TimeControl extends Entity {}
+
+type TimeTable = Map<string, () => Promise<void>>
+class TimeControl extends Entity { }
 const schema = new Schema(
     TimeControl,
     {
-        name: {type: "string"},
-        cooldown: {type: "date"}
+        name: { type: "string" },
+        cooldown: { type: "date" }
     },
     {
         dataStructure: "JSON"
@@ -22,7 +23,7 @@ const schema = new Schema(
 export default {
     async generate(form: TimeControl) {
         await client.open(process.env.REDIS_URL)
-        
+
         const repository: Repository<TimeControl> = client.fetchRepository(schema)
         const timeControl: TimeControl = repository.createEntity()
 
@@ -42,28 +43,63 @@ export default {
         const timeControlId: string = await repository.search().where("name").equals(name).return.firstId()
 
         await repository.remove(timeControlId)
-        
+
         await client.close()
     },
-    async check(table: TimeTable) {
+    async check(name: string, handler?: () => Promise<void>) {
         console.log("Checking time controls!")
         await client.open(process.env.REDIS_URL)
 
         const repository: Repository<TimeControl> = client.fetchRepository(schema)
         await repository.createIndex()
 
-        const all = await repository.search().all()
+        const timeControl = await repository.search().where("name").equals(name).return.first()
 
-        for (const timeControl of all) {
-            if (utils.time.past(timeControl.cooldown)) {
-                console.log("Cooldown met: " + timeControl.name + "!")
-                const whenRemove = table.get(timeControl.name)
-                if (whenRemove) await whenRemove()
-                await repository.remove(timeControl.entityId)
-            }
-            else {
-                console.log("Cooldown of " + timeControl.name + "not met...")
-            }
+        if (utils.time.past(timeControl.cooldown)) {
+            console.log("Cooldown met: " + timeControl.name + "!")
+            if (handler) await handler()
+
+            await repository.remove(timeControl.entityId)
+        }
+        else {
+            console.log("Cooldown of " + timeControl.name + " not met...")
+        }
+
+        await client.close()
+    },
+
+    async resume(table: TimeTable) {
+        await client.open(process.env.REDIS_URL)
+
+        const repository: Repository<TimeControl> = client.fetchRepository(schema)
+        await repository.createIndex()
+
+        const leftovers = await repository.search().all()
+        if (leftovers.length > 0) {
+        console.log("There's some unresolved time controls...")
+        const interval = setInterval(async () => {
+                const timeControls = await repository.search().all()
+                if (timeControls.length === 0) {
+                    console.log("That's a wrap!")
+                    clearInterval(interval)
+                    return
+                }
+                for (const timeControl of timeControls) {
+                    if (utils.time.past(timeControl.cooldown)) {
+                        console.log("Cooldown met: " + timeControl.name + "!")
+                        const handler = table.get(timeControl.name)
+
+                        if (handler) await handler()
+                        await repository.remove(timeControl.entityId)
+                    }
+                    else {
+                        console.log("Cooldown of " + timeControl.name + "not met...")
+                    }
+                }
+            }, 1000)
+        }
+        else {
+            console.log("Time controls are clear.")
         }
     }
 }
